refactor(core): split series sync into focused modules
Extract shared sync helpers (clampRating, fetchParallel, streamingBatchProcess)
and decompose the monolithic series/sync.ts into batch, reconcile, library,
and watch-history modules while preserving the public barrel exports.
